component refactoring
